Validations

+++ src/main/java/com/metrobank/uploadITR/service/Upload.java
@@ -1,5 +1,8 @@
 package com.metrobank.uploadITR.service;
 
+import com.metrobank.uploadITR.exception.FilePathValidationException;
+import com.metrobank.uploadITR.exception.ItrIdValidationException;
+import com.metrobank.uploadITR.exception.UserIdValidationException;
 import com.metrobank.uploadITR.model.UploadModel;
 import com.metrobank.uploadITR.repository.UploadRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,7 +12,7 @@ import java.util.List;
 
 @Service
 public class Upload {
-    private UploadRepository uploadRepository;
+    private final UploadRepository uploadRepository;
     @Autowired
     public Upload(UploadRepository uploadRepository){
         this.uploadRepository = uploadRepository;
@@ -17,6 +20,18 @@ public class Upload {
 
     //Upload Backend Logic
     public UploadModel upload (int user_id, int year, String file_path){
+
+        //No validation yet for unfilled user id and year
+        if(uploadRepository.existByUserId(user_id) == 0){
+            throw new UserIdValidationException("This user does not exist.");
+        }
+        if(file_path.isEmpty()){
+            throw new FilePathValidationException("File path must not be empty.");
+        }
+        if(!file_path.contains(".pdf")){
+            throw new FilePathValidationException("This is not a .pdf file.");
+        }
+
         UploadModel model = new UploadModel();
         model.setUserId(user_id);
         model.setYear(year);
@@ -26,6 +41,11 @@ public class Upload {
 
     //For removing an itr record
     public void remove (int itr_id){
+
+        if(uploadRepository.existByItrId(itr_id) == 0){
+            throw new ItrIdValidationException("This ITR record does not exist.");
+        }
+
         uploadRepository.deleteItrById(itr_id);
     }
